Landscape: Fix truncation warnings, part 2
#jira UE-166471
#preflight 6359e871316cf8bef619d6b9
#rb jonathan.hard

[CL 22809417 by michael balzer in ue5-main branch]

+++ Engine/Source/Runtime/Landscape/Private/LandscapeSubsystem.cpp
@@ -233,7 +233,7 @@ void ULandscapeSubsystem::BuildNanite()
 		return;
 	}
 
-	FScopedSlowTask SlowTask(Proxies.Num(), (LOCTEXT("Landscape_BuildNanite", "Building Nanite Landscape Meshes")));
+	FScopedSlowTask SlowTask(static_cast<float>(Proxies.Num()), (LOCTEXT("Landscape_BuildNanite", "Building Nanite Landscape Meshes")));
 	SlowTask.MakeDialog();
 
 	for (TWeakObjectPtr<ALandscapeProxy> ProxyPtr : Proxies)
